+++ packages/pwa-kit-create-app/assets/pwa/sites.js
@@ -1,0 +1,25 @@
+/*
+ * Copyright (c) 2021, salesforce.com, inc.
+ * All rights reserved.
+ * SPDX-License-Identifier: BSD-3-Clause
+ * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+exports.template = () => `// Provide the sites for your app. Each site includes site id, and its localization configuration.
+// You can also provide aliases for your locale. They will be used in place of your locale id when generating paths across the app
+module.exports = [
+    {
+        id: 'RefArch',
+        l10n: {
+            supportedCurrencies: ['USD'],
+            defaultCurrency: 'USD',
+            defaultLocale: 'en-US',
+            supportedLocales: [
+                {
+                    id: 'en-US',
+                    // alias: 'us',
+                    preferredCurrency: 'USD'
+                }
+            ]
+        }
+    }
+]`