+++ packages/pwa/app/utils/url.test.js
@@ -10,22 +10,35 @@ import {
     categoryUrlBuilder,
     productUrlBuilder,
     searchUrlBuilder,
-    getUrlWithLocale,
+    getPathWithLocale,
     homeUrlBuilder,
     rebuildPathWithParams,
     removeQueryParamsFromPath,
-    buildPathWithUrlConfig
+    buildPathWithUrlConfig,
+    absoluteUrl
 } from './url'
 import {getUrlConfig} from './utils'
+import mockConfig from '../../config/mocks/default'
 
+afterEach(() => {
+    jest.clearAllMocks()
+})
+
+jest.mock('pwa-kit-react-sdk/utils/url', () => {
+    const original = jest.requireActual('pwa-kit-react-sdk/utils/url')
+    return {
+        ...original,
+        getAppOrigin: jest.fn(() => 'https://www.example.com')
+    }
+})
 jest.mock('./utils', () => {
     const original = jest.requireActual('./utils')
     return {
         ...original,
+        getConfig: jest.fn(() => mockConfig),
         getUrlConfig: jest.fn()
     }
 })
-import {DEFAULT_LOCALE} from '../constants'
 
 describe('buildUrlSet returns the expected set of urls', () => {
     test('when no values are passed in', () => {
@@ -109,88 +122,195 @@ describe('url builder test', () => {
         const url = categoryUrlBuilder({id: 'men'})
         expect(url).toEqual(`/category/men`)
     })
+})
 
-    test('homeUrlBuilder returns expect', () => {
-        getUrlConfig.mockImplementation(() => ({
-            locale: 'path'
-        }))
-        const url = homeUrlBuilder('/', 'fr-FR')
-        expect(url).toEqual(`/fr-FR/`)
+describe('getPathWithLocale', () => {
+    getUrlConfig.mockImplementation(() => mockConfig.app.url)
+    test('getPathWithLocale returns expected for PLP', () => {
+        const location = new URL('http://localhost:3000/uk/it-IT/category/newarrivals-womens')
 
-        const homeUrlDefaultLocale = homeUrlBuilder('/', DEFAULT_LOCALE)
-        expect(homeUrlDefaultLocale).toEqual(`/`)
+        const relativeUrl = getPathWithLocale('fr-FR', {location})
+        expect(relativeUrl).toEqual(`/uk/fr-FR/category/newarrivals-womens`)
     })
 
-    test('getUrlWithLocale returns expected for PLP', () => {
-        getUrlConfig.mockImplementation(() => ({
-            locale: 'path'
-        }))
-        const location = new URL('http://localhost:3000/it-IT/category/newarrivals-womens')
+    test('getPathWithLocale uses default site for siteRef when it is no defined in the url', () => {
+        const location = new URL('http://localhost:3000/category/newarrivals-womens')
 
-        window.location = location
-
-        const relativeUrl = getUrlWithLocale('fr-FR')
-        expect(relativeUrl).toEqual(`/fr-FR/category/newarrivals-womens`)
+        const relativeUrl = getPathWithLocale('fr-FR', {location})
+        expect(relativeUrl).toEqual(`/uk/fr-FR/category/newarrivals-womens`)
     })
 
-    test('getUrlWithLocale returns expected for PLP without refine param', () => {
-        getUrlConfig.mockImplementation(() => ({
-            locale: 'path'
-        }))
+    test('getPathWithLocale returns expected for PLP without refine param', () => {
         const location = new URL(
-            'http://localhost:3000/it-IT/category/newarrivals-womens?limit=25&refine=c_refinementColor%3DBianco&sort=best-matches&offset=25'
+            'http://localhost:3000/uk/it-IT/category/newarrivals-womens?limit=25&refine=c_refinementColor%3DBianco&sort=best-matches&offset=25'
         )
 
-        window.location = location
-
-        const relativeUrl = getUrlWithLocale('fr-FR', {
-            disallowParams: ['refine']
+        const relativeUrl = getPathWithLocale('fr-FR', {
+            disallowParams: ['refine'],
+            location
         })
         expect(relativeUrl).toEqual(
-            `/fr-FR/category/newarrivals-womens?limit=25&sort=best-matches&offset=25`
+            `/uk/fr-FR/category/newarrivals-womens?limit=25&sort=best-matches&offset=25`
         )
     })
 
-    test('getUrlWithLocale returns expected for PLP', () => {
-        getUrlConfig.mockImplementation(() => ({
-            locale: 'path'
-        }))
-        const location = new URL('http://localhost:3000/it-IT/category/newarrivals-womens')
+    test('getPathWithLocale returns expected for Homepage', () => {
+        const location = new URL('http://localhost:3000/uk/it-IT/')
 
-        window.location = location
-
-        const relativeUrl = getUrlWithLocale('fr-FR')
-        expect(relativeUrl).toEqual(`/fr-FR/category/newarrivals-womens`)
+        const relativeUrl = getPathWithLocale('fr-FR', {location})
+        expect(relativeUrl).toEqual(`/uk/fr-FR/`)
     })
 
-    test('getUrlWithLocale returns expected for Homepage', () => {
-        getUrlConfig.mockImplementation(() => ({
-            locale: 'path'
-        }))
-        const location = new URL('http://localhost:3000/it-IT/')
+    test('getPathWithLocale returns / when both site and locale are default', () => {
+        const location = new URL('http://localhost:3000/uk/it-IT/')
 
-        window.location = location
-
-        const relativeUrl = getUrlWithLocale('fr-FR')
-        expect(relativeUrl).toEqual(`/fr-FR`)
-    })
-
-    test('getUrlWithLocale returns expected for Homepage with Default locale', () => {
-        const location = new URL('http://localhost:3000/it-IT/')
-
-        window.location = location
-
-        const relativeUrl = getUrlWithLocale(DEFAULT_LOCALE)
+        const relativeUrl = getPathWithLocale('en-GB', {location})
         expect(relativeUrl).toEqual(`/`)
     })
+})
 
-    test('getUrlWithLocale returns expected for Homepage without trailing slash', () => {
-        const location = new URL('http://localhost:3000/it-IT')
+describe('homeUrlBuilder', () => {
+    const defaultSite = mockConfig.app.sites[0]
+    const defaultAlias = mockConfig.app.siteAliases[defaultSite.id]
+    const defaultSiteMock = {...defaultSite, alias: defaultAlias}
 
-        window.location = location
+    const nonDefaultSite = mockConfig.app.sites[1]
+    const nonDefaultAlias = mockConfig.app.siteAliases[nonDefaultSite.id]
+    const nonDefaultSiteMock = {...nonDefaultSite, alias: nonDefaultAlias}
+    const cases = [
+        {
+            urlConfig: {
+                locale: 'path',
+                site: 'path',
+                showDefaults: true
+            },
+            site: defaultSiteMock,
+            locale: {id: 'en-GB'},
+            expectedRes: '/'
+        },
+        {
+            urlConfig: {
+                locale: 'query_param',
+                site: 'query_param',
+                showDefaults: true
+            },
+            site: defaultSiteMock,
+            locale: {id: 'en-GB'},
+            expectedRes: '/'
+        },
+        {
+            urlConfig: {
+                locale: 'path',
+                site: 'path',
+                showDefaults: false
+            },
+            site: defaultSiteMock,
+            locale: {id: 'en-GB'},
+            expectedRes: '/'
+        },
+        {
+            urlConfig: {
+                locale: 'query_param',
+                site: 'query_param',
+                showDefaults: false
+            },
+            site: defaultSiteMock,
+            locale: {id: 'en-GB'},
+            expectedRes: '/'
+        },
+        {
+            urlConfig: {
+                locale: 'path',
+                site: 'path',
+                showDefaults: true
+            },
+            site: defaultSiteMock,
+            locale: {id: 'fr-FR'},
+            expectedRes: '/uk/fr-FR/'
+        },
+        {
+            urlConfig: {
+                locale: 'path',
+                site: 'path',
+                showDefaults: false
+            },
+            site: defaultSiteMock,
+            locale: {id: 'fr-FR'},
+            expectedRes: '/fr-FR/'
+        },
+        {
+            urlConfig: {
+                locale: 'query_param',
+                site: 'query_param',
+                showDefaults: true
+            },
+            site: defaultSiteMock,
+            locale: {id: 'fr-FR'},
+            expectedRes: '/?site=uk&locale=fr-FR'
+        },
+        {
+            urlConfig: {
+                locale: 'path',
+                site: 'path',
+                showDefaults: true
+            },
+            site: nonDefaultSiteMock,
+            locale: {id: 'en-US'},
+            expectedRes: '/us/en-US/'
+        },
+        {
+            urlConfig: {
+                locale: 'query_param',
+                site: 'path',
+                showDefaults: true
+            },
+            site: nonDefaultSiteMock,
+            locale: {id: 'en-US'},
+            expectedRes: '/us/?locale=en-US'
+        },
+        {
+            urlConfig: {
+                locale: 'path',
+                site: 'path',
+                showDefaults: false
+            },
+            site: nonDefaultSiteMock,
+            locale: {id: 'en-US'}, // default locale of the nonDefault Site
+            expectedRes: '/us/'
+        },
+        {
+            urlConfig: {
+                locale: 'query_param',
+                site: 'path',
+                showDefaults: false
+            },
+            site: nonDefaultSiteMock,
+            locale: {id: 'en-US'}, // default locale of the nonDefault Site
+            expectedRes: '/us/'
+        },
+        {
+            urlConfig: {
+                locale: 'query_param',
+                site: 'query_param',
+                showDefaults: true
+            },
+            site: nonDefaultSiteMock,
+            locale: {id: 'en-US'}, // default locale of the nonDefault Site
+            expectedRes: '/?site=us&locale=en-US'
+        }
+    ]
 
-        const relativeUrl = getUrlWithLocale(DEFAULT_LOCALE)
-        expect(relativeUrl).toEqual(`/`)
+    cases.forEach(({urlConfig, site, locale, expectedRes}) => {
+        test(`return expected URL with site ${site.alias}, locale ${
+            locale.id
+        } and urlConfig as ${JSON.stringify(urlConfig)}`, () => {
+            getUrlConfig.mockImplementation(() => urlConfig)
+            const homeUrl = homeUrlBuilder('/', {
+                site,
+                locale
+            })
+            expect(homeUrl).toEqual(expectedRes)
+        })
     })
 })
 
@@ -211,21 +331,101 @@ describe('removeQueryParamsFromPath test', () => {
 })
 
 describe('buildPathWithUrlConfig', () => {
-    test('return a new url with locale value as a part of path', () => {
+    test('return a new url with locale and site a part of path', () => {
         getUrlConfig.mockImplementation(() => ({
-            locale: 'path'
+            locale: 'path',
+            site: 'path',
+            showDefaults: true
         }))
-
-        const url = buildPathWithUrlConfig('/women/dresses', {locale: 'en-GB'})
-        expect(url).toEqual('/en-GB/women/dresses')
+        const url = buildPathWithUrlConfig('/women/dresses', {locale: 'en-GB', site: 'uk'})
+        expect(url).toEqual('/uk/en-GB/women/dresses')
     })
 
-    test('return a new url with locale value as a query param', () => {
+    test('return an expected url with no site, no locale for default values when the showDefaults is off', () => {
         getUrlConfig.mockImplementation(() => ({
-            locale: 'query_param'
+            locale: 'path',
+            site: 'path',
+            showDefaults: false
         }))
+        const url = buildPathWithUrlConfig('/women/dresses', {locale: 'en-GB', site: 'uk'})
+        expect(url).toEqual('/women/dresses')
+    })
 
-        const url = buildPathWithUrlConfig('/women/dresses', {locale: 'en-GB'})
-        expect(url).toEqual('/women/dresses?locale=en-GB')
+    test('return a new url with locale value as a query param and site in the path', () => {
+        getUrlConfig.mockImplementation(() => ({
+            locale: 'query_param',
+            site: 'path',
+            showDefaults: true
+        }))
+        const url = buildPathWithUrlConfig('/women/dresses', {locale: 'en-GB', site: 'uk'})
+        expect(url).toEqual('/uk/women/dresses?locale=en-GB')
+    })
+
+    test('return a new url with locale value as a path, site as query_param', () => {
+        getUrlConfig.mockImplementation(() => ({
+            locale: 'path',
+            site: 'query_param',
+            showDefaults: true
+        }))
+        const url = buildPathWithUrlConfig('/women/dresses', {locale: 'en-GB', site: 'uk'})
+        expect(url).toEqual('/en-GB/women/dresses?site=uk')
+    })
+
+    test('return a new url with locale value as a path, site as query_param when showDefault is off', () => {
+        getUrlConfig.mockImplementation(() => ({
+            locale: 'path',
+            site: 'query_param',
+            showDefaults: false
+        }))
+        const url = buildPathWithUrlConfig('/women/dresses', {locale: 'en-GB', site: 'uk'})
+        expect(url).toEqual('/women/dresses')
+    })
+
+    test('return a new url without a disallow param but respect other params', () => {
+        getUrlConfig.mockImplementation(() => ({
+            locale: 'query_param',
+            site: 'path',
+            showDefaults: true
+        }))
+        const url = buildPathWithUrlConfig(
+            '/women/dresses?something=else&refine=c_color',
+            {locale: 'en-GB', site: 'uk'},
+            {disallowParams: ['refine']}
+        )
+        expect(url).toEqual('/uk/women/dresses?something=else&locale=en-GB')
+    })
+
+    test('return a new url as configured when the values are not defaults and showDefault is off', () => {
+        getUrlConfig.mockImplementation(() => ({
+            locale: 'query_param',
+            site: 'path',
+            showDefaults: false
+        }))
+        const url = buildPathWithUrlConfig(
+            '/women/dresses?something=else&refine=c_color',
+            {locale: 'en-CA', site: 'us'},
+            {disallowParams: ['refine']}
+        )
+        expect(url).toEqual('/us/women/dresses?something=else&locale=en-CA')
+    })
+
+    test('throw an error when url config is not defined', () => {
+        getUrlConfig.mockImplementation(() => undefined)
+
+        expect(() => {
+            buildPathWithUrlConfig('/women/dresses', {locale: 'en-GB', site: 'uk'})
+        }).toThrow()
+    })
+})
+
+describe('absoluteUrl', function() {
+    test('return expected when path is a relative url', () => {
+        const url = absoluteUrl('/uk/en/women/dresses')
+        expect(url).toEqual('https://www.example.com/uk/en/women/dresses')
+    })
+
+    test('return expected when path is an absolute url', () => {
+        const url = absoluteUrl('https://www.example.com/uk/en/women/dresses')
+        expect(url).toEqual('https://www.example.com/uk/en/women/dresses')
     })
 })