+++ packages/pwa/jest-setup.js
@@ -7,7 +7,7 @@
 import 'raf/polyfill' // fix requestAnimationFrame issue with polyfill
 import '@testing-library/jest-dom/extend-expect'
 import {Crypto} from '@peculiar/webcrypto'
-import config from './config/default'
+import mockConfig from './config/mocks/default'
 import {setConfig} from 'pwa-kit-react-sdk/ssr/universal/utils'
 
 // TextEncoder is a web API, need to import it
@@ -48,11 +48,13 @@ Object.defineProperty(window, 'scrollTo', {
     value: () => null
 })
 
-// Mock the the `default` config to the window global
+// Mock the `default` config to the window global
 Object.defineProperty(window, '__CONFIG__', {
-    value: config
+    value: mockConfig,
+    configurable: true
 })
-setConfig(config)
+
+setConfig(mockConfig)
 
 Object.defineProperty(window, 'matchMedia', {
     writable: true,